[Client] Separate webpack into dev and prod configs
