+++ package.json
@@ -5,9 +5,11 @@
     "start-server": "node server/dist/index.js",
     "start-server-autoreload": "nodemon server/dist/index.js -w server/dist",
     "dev-server": "npm-run-all -p watch-server-js start-server-autoreload",
-    "build-client-js": "webpack --config client/webpack.config.js",
-    "watch-client-js": "yarn build-client-js -- -w",
-    "dev": "npm-run-all -p dev-server watch-client-js"
+    "build-client-dev": "webpack --config client/webpack.dev.js",
+    "build": "webpack --config client/webpack.prod.js",
+    "watch-client-js": "yarn build-client-dev -- -w",
+    "dev": "npm-run-all -p dev-server watch-client-js",
+    "analyze": "webpack-bundle-analyzer ./client/dist/stats.json"
   },
   "devDependencies": {
     "@date-io/moment": "^1.0.1",
@@ -15,6 +17,7 @@
     "@material-ui/icons": "^3.0.1",
     "@types/auth0-js": "^8.11.6",
     "@types/bcryptjs": "^2.4.1",
+    "@types/compression": "^0.0.36",
     "@types/cors": "^2.8.3",
     "@types/debug": "^0.0.31",
     "@types/express": "^4.11.1",
@@ -24,10 +27,12 @@
     "@types/react-router-dom": "^4.3.1",
     "@types/sequelize": "^4.27.10",
     "auth0-js": "^9.8.1",
+    "clean-webpack-plugin": "^1.0.1",
     "copyfiles": "^2.1.0",
     "eslint": "^5.7.0",
     "file-loader": "^3.0.1",
     "fork-ts-checker-webpack-plugin": "^0.5.2",
+    "html-webpack-plugin": "^3.2.0",
     "material-ui-pickers": "^2.1.1",
     "moment": "^2.22.2",
     "nodemon": "^1.18.4",
@@ -39,11 +44,14 @@
     "ts-loader": "^5.3.2",
     "typescript": "^3.1.3",
     "webpack": "^4.22.0",
-    "webpack-cli": "^3.1.2"
+    "webpack-bundle-analyzer": "^3.0.3",
+    "webpack-cli": "^3.1.2",
+    "webpack-merge": "^4.2.1"
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.18.2",
+    "compression": "^1.7.3",
     "connect-session-sequelize": "^6.0.0",
     "cors": "^2.8.4",
     "express": "^4.16.2",
